Separate the parseConditionalExpr function and adapt the parsing function's pattern
